move http listener startup out of adapter
